+++ client/src/components/Onboarding/OnboardingActions.module.scss
@@ -1,0 +1,8 @@
+
+.onboarding-actions,
+.onboarding-processing {
+  .loading-icon {
+    display: inline-block;
+    width: inherit;
+  }
+}
