+++ client/src/components/Onboarding/OnboardingStatus.module.scss
@@ -1,0 +1,8 @@
+
+.onboarding-actions,
+.onboarding-processing {
+  .loading-icon {
+    display: inline-block;
+    width: inherit;
+  }
+}
