+++ client/src/components/Onboarding/OnboardingStep.module.scss
@@ -1,0 +1,15 @@
+
+.onboarding-step {
+  padding-top: 1em;
+  padding-bottom: 1em;
+  margin-bottom: 1em;
+  border-bottom: 1px solid var(--global-color-primary--normal);
+}
+
+.onboarding-step__name {
+  font-weight: bold;
+}
+
+.onboarding-step__description {
+  padding-bottom: 1em;
+}
