+++ client/src/components/Onboarding/OnboardingUser.scss
@@ -25,18 +25,3 @@
 .onboarding__container {
   width: 75%;
 }
-
-.onboarding-step {
-  padding-top: 1em;
-  padding-bottom: 1em;
-  margin-bottom: 1em;
-  border-bottom: 1px solid var(--global-color-primary--normal);
-}
-
-.onboarding-step__name {
-  font-weight: bold;
-}
-
-.onboarding-step__description {
-  padding-bottom: 1em;
-}
